refactor: JpaRepository infra계층으로 이동
