+++ src/main/java/kitchenpos/eatinorders/infra/JpaOrderRepository.java
@@ -1,8 +1,11 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.infra;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.eatinorders.domain.Order;
+import kitchenpos.eatinorders.domain.OrderRepository;
 
 public interface JpaOrderRepository extends OrderRepository, JpaRepository<Order, UUID> {
 }
