+++ src/main/java/kitchenpos/eatinorders/infra/JpaOrderTableRepository.java
@@ -1,8 +1,11 @@
-package kitchenpos.eatinorders.domain;
+package kitchenpos.eatinorders.infra;
 
 import java.util.UUID;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import kitchenpos.eatinorders.domain.OrderTable;
+import kitchenpos.eatinorders.domain.OrderTableRepository;
 
 public interface JpaOrderTableRepository extends OrderTableRepository, JpaRepository<OrderTable, UUID> {
 }
