+++ src/main/java/kitchenpos/menus/infra/JpaMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.menus.domain;
+package kitchenpos.menus.infra;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -6,6 +6,9 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 import java.util.UUID;
+
+import kitchenpos.menus.domain.Menu;
+import kitchenpos.menus.domain.MenuRepository;
 
 public interface JpaMenuRepository extends MenuRepository, JpaRepository<Menu, UUID> {
     @Query("select m from Menu m join m.menuProducts.menuProducts mp where mp.productId = :productId")
